build: refactor gradlew to set JVM options later in script
The change moves DEFAULT_JVM_OPTS setting further down in the script. This was done to ensure that JVM options are set later during the execution, after performing important configuration checks, rather than at the very beginning. This is to ensure that the configuration options are properly initialized and finalized before JVM options are set. It eliminates the chances of potential configuration errors."