build: refactor gradle build to use providers API and cleanup redundant code
This changes the build.gradle.kts script to implement the Gradle Providers API for a more idiosyncratic way of handling optional project properties. It also cleans up unnecessary code, making the script more efficient.

The use of providers API allows Gradle to be more lazy, only executing code when it is necessary. In some specific cases, this can significantly speed up the build. Removing excess code snippets that do not contribute to overall functionality also contributes to a cleaner and more efficient codebase.