ci: add Java setup to build workflow
The build.yml file for the GitHub Actions workflow was updated to include a step for setting up the Java environment. This change was necessary to ensure that Java 17 (Zulu distribution) is available and correctly set for the subsequent workflow steps. The Java setup step was added to three sections of the workflow: the Gradle Wrapper Validation, Fetch Sources, and Remove Old Release Drafts.