ci: add `test` and `inspectCode` workflow steps to build
This commit adds two important steps to Github Actions workflow. Firstly, it introduces a 'test' step, running unit tests and uploading a code coverage report. This makes sure that every commit or pull request is tested and potential issues are caught earlier. Secondly, the 'inspectCode' step is added, running Qodana inspections on the codebase which helps maintaining the code quality. Lastly, the needs of 'verify' and 'releaseDraft' workflows were updated to ensure the correct order of tasks.